Review changes: Move save logic out of the View
