+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/StateContainer.cs
@@ -33,8 +33,10 @@ namespace UnityEngine.InputSystem.Editor
 
             m_State = command(m_State);
 
+            // TODO: Need to filter out Save and AutoSave events as they should not trigger duplicate saves here
+            // Possibly they also shouldn't trigger the StateChanged callbacks either.
             if (InputEditorUserSettings.autoSaveInputActionAssets)
-                m_State.Save();
+                InputActionsEditorWindow.SaveAsset(m_State.serializedObject);
 
             // why not just invoke the state changed event immediately you ask? The Dispatch method might have
             // been called from inside a UI element event handler and if we raised the event immediately, a view
